+++ examples/glEvalLimit/glEvalLimit.cpp
@@ -60,7 +60,7 @@ GLFWmonitor* g_primary=0;
 #include <common/vtr_utils.h>
 #include "../common/stopwatch.h"
 #include "../common/simple_math.h"
-#include "../common/gl_hud.h"
+#include "../common/glHud.h"
 
 #include "init_shapes.h"
 #include "particles.h"
